+++ laravel/documentation/changes.md
@@ -64,6 +64,7 @@
 - Added `array_pluck` helper, similar to pluck method in Underscore.js.
 - Allow the registration of custom cache and session drivers.
 - Allow the specification of a separate asset base URL for using CDNs.
+- Allow a `starter` Closure to be defined in `bundles.php` to be run on Bundle::start.
 
 <a name="upgrade-3.2"></a>
 ## Upgrading From 3.1